Setup Project
- Cleared init files.
- Created folder structure.
