+++ src/app/App.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function App() {
+  return <div className="App">app</div>;
+}
+
+export default App;
